The WordPress plugin allows you to calculate the amortization schedule of a loan or a mortgage with the French method
Plugin make a button in the TinyMCE editor. Click button and calculate
mortgage plugin insert a shortcode "[calculate]" in you content. This
shortcode show the form for the calculation in your page or post. It
work only in euro and independently with dot or comma and accept only
numbers. Enter the required values ​​and press the botton for calculate
your loan.
Attention: The calculations have been tested many times but are subject
to approximations and are indicative. kiwi adv assumes no
responsibility for any use made with this plugin.